CHG move command manager to its own project
