--- src/Executables/Game/Commands/CommandDescriptor.cs
@@ -1,4 +1,0 @@
-﻿namespace QuantumCore.Game.Commands;
-
-internal record CommandDescriptor(Type Type, string Command, string? Description = null,
-    Type? OptionsType = null, bool BypassPerm = false);
