+++ src/Executables/Game/Commands/HelpCommand.cs
@@ -1,4 +1,5 @@
-﻿using QuantumCore.API.Game;
+﻿using QuantumCore.API;
+using QuantumCore.API.Game;
 using QuantumCore.API.Game.World;
 
 namespace QuantumCore.Game.Commands
