+++ src/Executables/Game/PlayerUtils/ExperienceManager.cs
@@ -1,5 +1,6 @@
 ﻿using System.Diagnostics;
 using Microsoft.Extensions.Logging;
+using QuantumCore.API;
 
 namespace QuantumCore.Game.PlayerUtils;
