+++ src/Executables/Game/PlayerUtils/JobManager.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
+using QuantumCore.API;
 using QuantumCore.API.Game.Types;
 
 namespace QuantumCore.Game.PlayerUtils
@@ -63,7 +64,7 @@ namespace QuantumCore.Game.PlayerUtils
             _logger.LogError("Invalid status {Status}", str);
             return EPoints.St;
         }
-        
+
         public Job Get(byte playerClass)
         {
             var index = GetJobFromClass(playerClass);
