+++ src/Libraries/Game.Commands/CommandContext.cs
@@ -1,0 +1,5 @@
+﻿using QuantumCore.API.Game.World;
+
+public record struct CommandContext<T>(IPlayerEntity Player, T Arguments);
+
+public record struct CommandContext(IPlayerEntity Player);
