+++ src/Libraries/Game.Commands/CommandDescriptor.cs
@@ -1,0 +1,8 @@
+﻿namespace QuantumCore.Game.Commands;
+
+internal record CommandDescriptor(
+    Type Type,
+    string Command,
+    string? Description = null,
+    Type? OptionsType = null,
+    bool BypassPerm = false);
