+++ src/Libraries/Game.Commands/ICommandHandler.cs
@@ -1,0 +1,11 @@
+﻿namespace QuantumCore.Game.Commands;
+
+public interface ICommandHandler<T>
+{
+    Task ExecuteAsync(CommandContext<T> context);
+}
+
+public interface ICommandHandler
+{
+    Task ExecuteAsync(CommandContext context);
+}
