+++ src/QuantumCore.sln
@@ -48,6 +48,12 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Game.Tests", "Tests\Game.Te
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Auth.Tests", "Tests\Auth.Tests\Auth.Tests.csproj", "{5C5AC92D-7010-4C24-A94D-7922787A58FE}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Libraries", "Libraries", "{9DA1A834-39B0-424F-A527-C0E9B2CBD931}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Game.Commands", "Libraries\Game.Commands\Game.Commands.csproj", "{493DD2DA-A3B3-4122-8DB4-7C854F80EB7D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Game.Commands.Tests", "Tests\Game.Commands.Tests\Game.Commands.Tests.csproj", "{EFF23932-7926-4DDD-8B45-9E7E2A2CA03A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -138,6 +144,14 @@ Global
 		{5C5AC92D-7010-4C24-A94D-7922787A58FE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5C5AC92D-7010-4C24-A94D-7922787A58FE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5C5AC92D-7010-4C24-A94D-7922787A58FE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{493DD2DA-A3B3-4122-8DB4-7C854F80EB7D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{493DD2DA-A3B3-4122-8DB4-7C854F80EB7D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{493DD2DA-A3B3-4122-8DB4-7C854F80EB7D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{493DD2DA-A3B3-4122-8DB4-7C854F80EB7D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EFF23932-7926-4DDD-8B45-9E7E2A2CA03A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EFF23932-7926-4DDD-8B45-9E7E2A2CA03A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EFF23932-7926-4DDD-8B45-9E7E2A2CA03A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EFF23932-7926-4DDD-8B45-9E7E2A2CA03A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{930E5131-B018-4255-BC08-ACE8F2C1BAC5} = {6D101418-9A2D-41CB-BEA9-47F915395CE8}
@@ -154,5 +168,7 @@ Global
 		{D02F020D-2464-48F3-B89E-EF4BABCB7713} = {5BBCB9D9-C272-4400-ABCE-B47C073AF753}
 		{0051033D-1CCB-460C-9E3D-7027141B437A} = {5BBCB9D9-C272-4400-ABCE-B47C073AF753}
 		{5C5AC92D-7010-4C24-A94D-7922787A58FE} = {5BBCB9D9-C272-4400-ABCE-B47C073AF753}
+		{493DD2DA-A3B3-4122-8DB4-7C854F80EB7D} = {9DA1A834-39B0-424F-A527-C0E9B2CBD931}
+		{EFF23932-7926-4DDD-8B45-9E7E2A2CA03A} = {5BBCB9D9-C272-4400-ABCE-B47C073AF753}
 	EndGlobalSection
 EndGlobal
